--- rdb_type/__init__.py
@@ -1,4 +1,0 @@
-""" Import all rdb type classes
-"""
-from .mysql import *
-from .synapse_database import *
