+++ rdb_updater/__init__.py
@@ -1,0 +1,3 @@
+""" Import all rdb type classes
+"""
+from .rdb_updater import *
